# api/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KU-Detection-Back-End API",
    "version": "1.0.0",
    "description": "API for the KU-Detection-Back-End project."
  },
...
  "paths": {
    "/cluster_repos": {
      "post": {
        "tags": [
          "Clustering"
        ],
        "summary": "Cluster Repositories by KUs",
        "description": "Performs K-Means clustering on repositories based on their detected Knowledge Units (KUs). The number of clusters is specified in the request body.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "num_clusters"
                ],
                "properties": {
                  "num_clusters": {
                    "type": "integer",
                    "description": "The desired number of clusters to create. Must be 2 or greater.",
                    "example": 3,
                    "minimum": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Clustering performed successfully. Returns a list of repositories with their assigned cluster ID.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "repo_name": {
                        "type": "string",
                        "description": "The name of the repository.",
                        "example": "kafka-python"
                      },
                      "cluster": {
                        "type": "integer",
                        "description": "The ID of the cluster the repository belongs to.",
                        "example": 1
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. The 'num_clusters' parameter is missing, not an integer, less than 2, or there are not enough repositories to form the requested clusters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Not enough repositories with detected KUs to form the requested number of clusters."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An internal error occurred during the clustering process.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to perform clustering due to an internal error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/commits": {
      "post": {
        "tags": [
          "Repository Management"
        ],
        "summary": "Get Commits",
        "description": "Retrieves a list of commits from a specified repository.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "repo_url": {
                    "type": "string",
                    "description": "URL of the repository",
                    "example": "https://github.com/user/repo.git"
                  },
                  "limit": {
                    "type": "integer",
                    "description": "Maximum number of commits to retrieve",
                    "example": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "author": {
                        "type": "string",
                        "description": "Author of the commit"
                      },
                      "message": {
                        "type": "string",
                        "description": "Commit message"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/repos": {
      "post": {
        "tags": [
          "Repository Management"
        ],
        "summary": "Create Repository",
        "description": "Creates a new repository entry.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "repo_name": {
                    "type": "string",
                    "description": "Name of the repository",
                    "example": "my-repo"
                  },
                  "url": {
                    "type": "string",
                    "description": "URL of the repository",
                    "example": "https://github.com/user/my-repo.git"
                  },
                  "organization": {
                    "type": "string",
                    "description": "The organization or owner of the repository",
                    "example": "apache"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the repository",
                    "example": "My cool project"
                  },
                  "comments": {
                    "type": "string",
                    "description": "Comments about the repository",
                    "example": "This repository is for..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Repository created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message",
                      "example": "Repository created successfully"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message",
                      "example": "Database connection failed"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Repository Management"
        ],
        "summary": "List Repositories",
        "description": "Lists all repository entries",
        "responses": {
          "200": {
            "description": "Successsful Operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "repo_name": {
                        "type": "string",
                        "description": "Name of repository"
                      },
                      "url": {
                        "type": "string",
                        "description": "Repository URL"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message",
                      "example": "Database connection failed"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/detected_kus": {
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "Get Detected KUs",
        "description": "Retrieves detected KUs.",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "ku_id": {
                        "type": "integer",
                        "description": "KU ID",
                        "example": 1
                      },
                      "ku_name": {
                        "type": "string",
                        "description": "KU Name",
                        "example": "Knowledge Unit 1"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message",
                      "example": "Database query failed"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/repos/{repo_name}": {
      "put": {
        "tags": [
          "Repository Management"
        ],
        "summary": "Edit Repository",
        "description": "Edits an existing repository entry.",
        "parameters": [
          {
            "name": "repo_name",
            "in": "path",
            "required": true,
            "description": "Name of the repository to edit",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "description": "URL of the repository",
                    "example": "https://github.com/user/my-repo.git"
                  },
                  "organization": {
                    "type": "string",
                    "description": "The organization or owner of the repository",
                    "example": "apache"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the repository",
                    "example": "My updated cool project"
                  },
                  "comments": {
                    "type": "string",
                    "description": "Comments about the repository",
                    "example": "These are the updated comments"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repository updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message",
                      "example": "Repository updated successfully"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message",
                      "example": "Database update failed"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/delete_repo/{repo_name}": {
      "delete": {
        "tags": [
          "Repository Management"
        ],
        "summary": "Delete Repository",
        "description": "Delete a repository entry",
        "parameters": [
          {
            "name": "repo_name",
            "in": "path",
            "required": true,
            "description": "Repo to Delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error Message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}